Removes unused old test harnesses and the unused GildedRose file, Changes the name of runner to GildedRose
